Updated File Service and Implementation

+++ pocketsend-backend/src/main/java/com/tomato/pocketsend/pocketsend_backend/service/FileServiceImpl.java
@@ -2,50 +2,55 @@ package com.tomato.pocketsend.pocketsend_backend.service;
 
 import com.tomato.pocketsend.pocketsend_backend.mappers.FileMapper;
 import com.tomato.pocketsend.pocketsend_backend.model.FileDTO;
+import com.tomato.pocketsend.pocketsend_backend.model.UserDTO;
 import com.tomato.pocketsend.pocketsend_backend.repositories.FileRepository;
-import com.tomato.pocketsend.pocketsend_backend.entity.File;
-import com.tomato.pocketsend.pocketsend_backend.entity.User;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class FileServiceImpl implements FileService{
-    @Autowired
-    private FileRepository fileRepository;
-    private FileMapper fileMapper;
 
-    public FileDTO saveFile(String filename, String filetype, byte[] content, User user) {
-        File file = new File();
-        file.setFilename(filename);
-        file.setFiletype(filetype);
-        file.setContent(content);
-        file.setUploadedAt(LocalDateTime.now());
-        file.setOwner(user);
+    private final FileRepository fileRepository;
+    private final FileMapper fileMapper;
 
-        return fileMapper.fileToFileDto(fileRepository.save(file));
+    @Override
+    public FileDTO saveFile(FileDTO file) {
+
+        return fileMapper.fileToFileDto(fileRepository
+                .save(fileMapper.fileDtoToFile(file)));
     }
 
+    @Override
     public List<FileDTO> getAllFiles() {
         return fileRepository.findAll().stream()
                 .map(fileMapper::fileToFileDto)
                 .collect(Collectors.toList());
     }
 
-    public List<File> getFilesForUser(User user) {
+    @Override
+    public List<FileDTO> getFilesForUser(UserDTO user) {
         return fileRepository.findByOwner(user);
     }
 
-    public Optional<File> getFileById(long id) {
-        return fileRepository.findById(id);
+    @Override
+    public Optional<FileDTO> getFileById(UUID id) {
+        return Optional.ofNullable(fileMapper.fileToFileDto(fileRepository.findById(id)
+                .orElse(null)));
     }
 
-    public void deleteFile(long id) {
-        fileRepository.deleteById(id);
+    @Override
+    public Boolean deleteFileById(UUID id) {
+        if(fileRepository.existsById(id)) {
+            fileRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
